Split CEED unit tests and main file for faster compilation
