+++ tests/unit/makefile
@@ -25,7 +25,8 @@ CCC = $(CC) $(CCOPTS)
 INCLUDES = -I$(or $(SRC:%/=%),.) -I$(MFEM_DIR)
 
 SOURCE_FILES := $(sort $(wildcard $(SRC)*/*.cpp))
-SOURCE_FILES := $(filter-out ceed/test_ceed.cpp, $(SOURCE_FILES))
+CEED_SOURCE_FILES = $(sort $(wildcard $(SRC)ceed/*.cpp))
+SOURCE_FILES := $(filter-out $(CEED_SOURCE_FILES), $(SOURCE_FILES))
 HEADER_FILES = $(SRC)catch.hpp $(SRC)unit_tests.hpp
 OBJECT_FILES = $(SOURCE_FILES:$(SRC)%.cpp=%.o)
 DATA_DIR = data
@@ -59,8 +60,7 @@ PAR_UNIT_TESTS = punit_tests $(PAR_SEDOV_TESTS)
 
 # Ceed tests
 USE_CEED := $(MFEM_USE_CEED:NO=)
-CEED_FILES = $(SRC)ceed/test_ceed.cpp
-CEED_MAIN_OBJ = $(SRC)ceed/test_ceed.o
+CEED_OBJ = $(CEED_SOURCE_FILES:$(SRC)%.cpp=%.o)
 SEQ_UNIT_TESTS += $(if $(USE_CEED),ceed_tests)
 
 ifeq ($(MFEM_USE_MPI),NO)
@@ -84,8 +84,8 @@ punit_tests: $(PAR_MAIN_OBJ) $(OBJECT_FILES) $(MFEM_LIB_FILE) $(CONFIG_MK) $(DAT
 cunit_tests: $(CUDA_MAIN_OBJ) $(OBJECT_FILES) $(MFEM_LIB_FILE) $(CONFIG_MK) $(DATA_DIR)
 	$(CCC) $(CUDA_MAIN_OBJ) $(OBJECT_FILES) $(MFEM_LINK_FLAGS) $(MFEM_LIBS) -o $(@)
 
-ceed_tests: $(CEED_MAIN_OBJ) $(MFEM_LIB_FILE) $(CONFIG_MK) $(DATA_DIR)
-	$(CCC) $(CEED_MAIN_OBJ) $(MFEM_LINK_FLAGS) $(MFEM_LIBS) -o $(@)
+ceed_tests: $(CEED_OBJ) $(MFEM_LIB_FILE) $(CONFIG_MK) $(DATA_DIR)
+	$(CCC) $(CEED_OBJ) $(MFEM_LINK_FLAGS) $(MFEM_LIBS) -o $(@)
 
 # Note: in this rule, we always use the full path to the source file as a
 # workaround for an issue with coveralls.
@@ -94,7 +94,7 @@ $(OBJECT_FILES) $(SEQ_MAIN_OBJ) $(PAR_MAIN_OBJ) $(CUDA_MAIN_OBJ): %.o: $(SRC)%.c
 	@mkdir -p $(@D)
 	$(CCC) -c $(abspath $(<)) $(MFEM_FLAGS) $(INCLUDES) -o $(@)
 
-$(CEED_MAIN_OBJ): $(CEED_FILES) $(HEADER_FILES) $(CONFIG_MK)
+$(CEED_OBJ): %.o: $(SRC)%.cpp $(HEADER_FILES) $(CONFIG_MK)
 	$(CCC) -c $(abspath $(<)) $(MFEM_FLAGS) $(INCLUDES) -o $(@)
 
 # Template rule for devices
